Fix tests after moving duplicate classes
