+++ subscribing-sdk/src/test/java/com/ably/tracking/subscriber/FactoryUnitTests.kt
@@ -3,6 +3,7 @@ package com.ably.tracking.subscriber
 import android.annotation.SuppressLint
 import android.content.Context
 import com.ably.tracking.AblyConfiguration
+import com.ably.tracking.BuilderConfigurationIncompleteException
 import com.ably.tracking.LogConfiguration
 import io.mockk.mockk
 import org.junit.Assert
